Add epoll driver and start io_uring (#3)
* Add some more term stuff and fcntl

* Fix some fcntl and non-block tests and doc

* Fix botched fmt

* Fix botched fmt

* Fix up process spawning bugs and add closures

* Add epoll, clean up new lints

* Fix unused warning, implement AsRawFd for file

* Add basic epoll to tiny-std

* fmt

* Cleanup uring skeleton

+++ rusl/src/platform/compat/socket.rs
@@ -72,8 +72,8 @@ transparent_bitflags!(
         const SOCK_SEQPACKET = 5;
         /// Deprecated
         const SOCK_PACKET = 10;
-        const SOCK_NONBLOCK = linux_rust_bindings::fcntl::O_NONBLOCK as i32;
-        const SOCK_CLOEXEC = linux_rust_bindings::fcntl::O_CLOEXEC as i32;
+        const SOCK_NONBLOCK = linux_rust_bindings::fcntl::O_NONBLOCK;
+        const SOCK_CLOEXEC = linux_rust_bindings::fcntl::O_CLOEXEC;
     }
 );
 
@@ -88,7 +88,6 @@ pub struct SocketArg {
 pub struct SocketAddress(linux_rust_bindings::socket::sockaddr_un);
 
 impl SocketAddress {
-
     /// Get the `AddressFamily` of the socket address
     #[inline]
     #[must_use]
